Agregar habitos a db y fix en fragment

+++ app/src/main/java/com/pruden/habits/clases/data/Habito.kt
@@ -5,7 +5,7 @@ data class Habito (
     val nombre: String,
     val listaValores: MutableList<Int>,
     val listaNotas: MutableList<String>,
-    val objetivo: Int,
+    val objetivo: Float,
     val tipoNumerico: Boolean,
     val unidad: String
 )
